refactor: Reservation 생성자 오버로딩 추가
