+++ src/main/java/roomescape/reservation/dto/ReservationReqDto.java
@@ -1,0 +1,13 @@
+package roomescape.reservation.dto;
+
+import roomescape.reservation.model.Reservation;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+public record ReservationReqDto(String name, LocalDate date, LocalTime time) {
+
+    public Reservation toEntity() {
+        return new Reservation(name, date, time);
+    }
+}
